Move token code to hashkit library
