+++ src/hashkit/dyn_token.c
@@ -23,7 +23,6 @@
 #include <stdint.h>
 #include <stddef.h>
 #include "dyn_token.h"
-#include "dyn_util.h"
 
 /* bitsPerDigit radix of 10 multiplied by 1024, rounded up to avoid underallocation */
 #define BITS_PER_DIGIT 3402
